+++ CHANGELOG.md
@@ -12,9 +12,12 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 - :fire: Remove public getters from classes. Now should be access via property instead.
 - :fire: Remove Pascal Case in properties. Properties updated to be camel case only.
 - :fire: Remove deprecated classes under `NelsonMartell\Utilities` namespace.
-- :bug: :fire: Fix problem in `NelsonMartell\StrictObject::compare()` for some types and make it more generic.
 - :new: `class`: `NelsonMartell\Extensions\Numbers`.
 - :new: `class`: `NelsonMartell\Extensions\Arrays`.
+- :bug: :fire: Fix problem in `NelsonMartell\StrictObject::compare()` for some types and make it more generic and split implementation to:
+  - :new: `NelsonMartell\Text::compare()`
+  - :new: `NelsonMartell\Arrays::compare()`
+  - :new: `NelsonMartell\Numbers::compare()`
 - :new: Add methods to check if a list of objects are of a type:
   - :new: `NelsonMartell\Type::is()`
   - :new: `NelsonMartell\Type::isIn()`